requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
redis>=5.0.0
pydantic-settings>=2.0.0
alembic>=1.12.0
